2004-10-12  Patrick Eriksson  <Patrick.Eriksson@rss.chalmers.se>
	* arts-1-1-900

	* Main things:
	Some cleaning and fixed so ppath_calc can be used from MC part.
	Some code and functions in MC are now obselete but left until changes
	are tested more carefully (testall.py runs fine!).

	Some old code removed in clear sky part.

	Code now compiles with Numeric=float. Clear sky part seems to
	work. Error when reading a scattering properties file. See email to
	arts-dev.

	Obvoius error in iy_interp_cloudbox_field fixed.

	* src/cloudbox.cc (iy_interp_cloudbox_field): Fixed problem pointed
	out by Cory. Typo made that scat_i_p was used instead of scat_i_lon.

	* src/rte.cc: Removed rte_calc_old and get_radiative_background_old.

	* src/m_rte.cc (RteEmissionStd): Mean of log(p) is now used (instead
	of mean of p).

 	* src/m_abs_o2_models.cc (CEF):
	src/m_atmosphere.cc (surfaceFlat):
	src/refraction.cc (refr_index_ir):

	Included type casting of fixed numeric values to avoid compilation
	warning with Numeric = float.

	* src/interpolation.cc: Changed FD_TOL from 1e-4 to 1e-3 to get
	Numeric=float to run.

	* src/ppath.cc (ppath_calc): Made necessary changes to allow
	that the function is also used for MC inside the cloud box.

	* src/m_montecarlo.cc
	(scat_iPutMonteCarlo): Replaced i_rte with
	iy, which is new variable with same purpose. This method is obselete
	but is left until things have been checked more carefully.

	(ScatteringMonteCarlo): Replaced call of Cloudbox_ppathCalc with
	call of ppath_calc.

	* src/montecarlo.cc (Cloudbox_ppath_rteCalc): Replaced call of
	Cloudbox_ppathCalc with call of ppath_calc.

	* src/workspace.cc: Removed surface_refl_coeffs and i_rte.

	* src/m_atmosphere.cc: Removed surfaceTreatAsBlackbody.

	* src/agendas.cc: Removed surface_agenda.

	* Untabified some files, but stop to avoid possibel merge problems
	if someone else does the same until I have a chance to commit.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@4276 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/m_abs_o2_models.cc
@@ -320,14 +320,21 @@ void CEF( // Input
   Complex retval;
 
   retval = (1.000e0 / sqrt(PI) /  Gamma_D) *
-             (  122.60793178 * pow(zeta,0.) 
-	      + 214.38238869 * pow(zeta,1.) + 181.92853309 * pow(zeta,2.) 
-	      +  93.15558046 * pow(zeta,3.) +  30.18014220 * pow(zeta,4.)
-	      +   5.91262621 * pow(zeta,5.) +   0.56418958 * pow(zeta,6.)) /
-             (  122.60793178 * pow(zeta,0.) + 352.73062511 * pow(zeta,1.)
-              + 457.33447878 * pow(zeta,2.) + 348.70391772 * pow(zeta,3.) 
-              + 170.35400182 * pow(zeta,4.) +  53.99290691 * pow(zeta,5.)
-              +  10.47985711 * pow(zeta,6.) +   1.00000000 * pow(zeta,7.));
+    (  122.60793178 * pow(zeta,Numeric(0.)) 
+       + 214.38238869 * pow(zeta,Numeric(1.)) 
+       + 181.92853309 * pow(zeta,Numeric(2.)) 
+       +  93.15558046 * pow(zeta,Numeric(3.)) 
+       +  30.18014220 * pow(zeta,Numeric(4.))
+       +   5.91262621 * pow(zeta,Numeric(5.)) 
+       +   0.56418958 * pow(zeta,Numeric(6.))) /
+    (  122.60793178 * pow(zeta,Numeric(0.)) 
+       + 352.73062511 * pow(zeta,Numeric(1.))
+       + 457.33447878 * pow(zeta,Numeric(2.)) 
+       + 348.70391772 * pow(zeta,Numeric(3.)) 
+       + 170.35400182 * pow(zeta,Numeric(4.)) 
+       +  53.99290691 * pow(zeta,Numeric(5.))
+       +  10.47985711 * pow(zeta,Numeric(6.)) 
+       +   1.00000000 * pow(zeta,Numeric(7.)));
     
   Re_CEF = real( retval );
   Im_CEF = imag( retval );
